# dagine/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "旨在学习和娱乐的社交软件",
        "title": "双鱼社交平台",
        "contact": {
            "name": "lucas",
            "url": "http://www.swagger.io/support",
            "email": "1783022886@qq.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8630",
    "basePath": "/lg_server/api",
    "paths": {
        "/health-check": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "健康检查"
                ],
                "summary": "健康检查",
                "responses": {}
            }
        },
        "/v1/demo/info": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "demo相关接口"
                ],
                "summary": "demo信息编辑",
                "parameters": [
                    {
                        "description": "demo信息",
                        "name": "update_info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    }
                }
            }
        },
        "/v1/demo/info/{demo_id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "demo相关接口"
                ],
                "summary": "查询demo详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "demoID",
                        "name": "demo_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/base.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "demo相关接口"
                ],
                "summary": "demo注销删除",
                "parameters": [
                    {
                        "type": "string",
                        "description": "demoID",
                        "name": "demo_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "0|1 登录|登出",
                        "name": "status",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    }
                }
            }
        },
        "/v1/demo/log_in_out/{demo_id}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "demo相关接口"
                ],
                "summary": "demo登录\u0026登出",
                "parameters": [
                    {
                        "type": "string",
                        "description": "demoID",
                        "name": "demo_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "0|1 登录|登出",
                        "name": "status",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    }
                }
            }
        },
        "/v1/demo/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "demo相关接口"
                ],
                "summary": "demo注册",
                "parameters": [
                    {
                        "description": "demo信息",
                        "name": "demo_info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "base.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "错误码",
                    "type": "integer"
                },
                "current_time": {
                    "description": "接口返回时间（当前时间）",
                    "type": "string"
                },
                "data": {
                    "description": "返回数据"
                },
                "err_trace": {
                    "description": "错误追踪链路信息",
                    "type": "string"
                },
                "message": {
                    "description": "错误信息",
                    "type": "string"
                },
                "version": {
                    "description": "版本信息",
                    "type": "string"
                }
            }
        }
    }
}